refactor: remove CreateTodoistSubtasksAsync from domain models
Complete separation of concerns by removing all Todoist-specific methods
from FoodServing hierarchy (FoodServing, CompositeFoodServing, StaticFoodServing).

Changes:
- Remove CreateTodoistSubtasksAsync() from FoodServing, CompositeFoodServing,
  and StaticFoodServing classes
- Add TodoistServiceHelper.CreateTodoistSubtasksAsync(FoodServing, ...)
  static method to handle all Todoist task creation
- Update TodoistService.AddServingAsync to use static helper
- Update CompositeFoodServingTests to use static helper
- Add SystemOfEquations.Todoist using statement to test file

Domain models are now completely free of Todoist-specific logic.
All Todoist integration is contained within the Todoist namespace.
