+++ Executable/FoodServing.cs
@@ -67,15 +67,4 @@ public record FoodServing(
     {
         return this * scale;
     }
-
-    // Virtual method to handle Todoist task creation
-    // Returns the created task ID if a parent task was created, null otherwise
-    public virtual async Task<string?> CreateTodoistSubtasksAsync(
-        string parentTaskId,
-        Func<string, string?, string?, string?, string?, Task<object>> addTaskFunc)
-    {
-        // Base FoodServing creates a single subtask
-        await addTaskFunc(ToString(), null, null, parentTaskId, null);
-        return null;
-    }
 }
